+++ cmd/plan.go
@@ -7,14 +7,15 @@ import (
 	"gopkg.in/nullstone-io/go-api-client.v0"
 	"gopkg.in/nullstone-io/go-api-client.v0/types"
 	"gopkg.in/nullstone-io/nullstone.v0/runs"
+	"os"
 	"strings"
 )
 
-var Up = func() *cli.Command {
+var Plan = func() *cli.Command {
 	return &cli.Command{
-		Name:      "up",
-		Usage:     "Provisions the block and all of its dependencies",
-		UsageText: "nullstone up [--stack=<stack-name>] --block=<block-name> --env=<env-name> [options]",
+		Name:      "plan",
+		Usage:     "Runs a plan with a disapproval",
+		UsageText: "nullstone plan [--stack=<stack-name>] --block=<block-name> --env=<env-name> [options]",
 		Flags: []cli.Flag{
 			StackFlag,
 			BlockFlag,
@@ -22,26 +23,31 @@ var Up = func() *cli.Command {
 			&cli.BoolFlag{
 				Name:    "wait",
 				Aliases: []string{"w"},
-				Usage:   "Wait for Nullstone to fully provision the workspace.",
+				Usage:   "Stream the Terraform logs while waiting for Nullstone to run the plan.",
 			},
 			&cli.StringSliceFlag{
 				Name:  "var",
-				Usage: "Set variable values when issuing `up`",
+				Usage: "Set variable values when issuing `plan`",
+			},
+			&cli.StringFlag{
+				Name:  "module-version",
+				Usage: "Use a specific module version to run the plan.",
 			},
 		},
 		Action: func(c *cli.Context) error {
 			varFlags := c.StringSlice("var")
+			moduleVersion := c.String("module-version")
 
 			return BlockWorkspaceAction(c, func(ctx context.Context, cfg api.Config, stack types.Stack, block types.Block, env types.Environment, workspace types.Workspace) error {
-				if workspace.Status == types.WorkspaceStatusProvisioned {
-					fmt.Println("workspace is already provisioned")
-					return nil
+				moduleSourceOverride := ""
+				if moduleVersion != "" {
+					moduleSourceOverride = fmt.Sprintf("%s@%s", block.ModuleSource, moduleVersion)
+				}
+				newRunConfig, err := runs.GetPromotion(cfg, workspace, moduleSourceOverride)
+				if err != nil {
+					return fmt.Errorf("error getting run configuration for plan: %w", err)
 				}
 
-				newRunConfig, err := runs.GetPromotion(cfg, workspace, "")
-				if err != nil {
-					return err
-				}
 				skipped, err := runs.SetRunConfigVars(newRunConfig, varFlags)
 				if len(skipped) > 0 {
 					fmt.Printf("[Warning] The following variables were skipped because they don't exist in the module: %s\n\n", strings.Join(skipped, ", "))
@@ -50,14 +56,14 @@ var Up = func() *cli.Command {
 					return err
 				}
 
-				t := true
-				newRun, err := runs.Create(cfg, workspace, newRunConfig, &t, false)
+				f := false
+				newRun, err := runs.Create(cfg, workspace, newRunConfig, &f, false)
 				if err != nil {
 					return fmt.Errorf("error creating run: %w", err)
 				} else if newRun == nil {
 					return fmt.Errorf("unable to create run")
 				}
-				fmt.Printf("created run %q\n", newRun.Uid)
+				fmt.Fprintf(os.Stdout, "created plan run %q\n", newRun.Uid)
 
 				if c.IsSet("wait") {
 					return runs.StreamLogs(ctx, cfg, workspace, newRun)